openssl: build for cygwin
Build for cygwin and patch the DLL names instead.
Also remove the OSNAME check in the perl scripts as I'm
not sure if perl will continue to report "msys" and we might
switch it to "cygwin" too.

While at it also re-enable makeflags, building with multiple
cores seems to work fine here, and is much faster.

All tests pass with this.