filesystem: move MAYBE_FIRST_START out of the post-install scripts
The main reason we have the first install warning still is that we
fill $HOME with bash related scripts that will only get executed
on the next login shell.

Simplify things by keeping MAYBE_FIRST_START logic in one file
and make it depend on wether $HOME was missing at the beginning.