less: Update to v621
There are many commits in the main branch of `less` since the last
version that was released "for general use" (v608).

Hoping that some of those commits fix issues reported in the Git for
Windows bug tracker, let's update to the latest tagged version, v621.

An alternative would be to backport individual patches, but due to some
sweeping changes (e.g. to convert from Kernighan & Ritchie style to C89
style function definitions), I've failed to backport them separately.

Switch to a GitHub archive since this release isn't available on
the website yet.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>