vim: avoid linking to the perl DLL
The major Perl version is encoded in the file name of the Perl DLL,
which makes it much more desirable to link _dynamically_ to the DLL,
i.e. on demand. Otherwise `vim.exe` would not even start unless Perl is
installed, at the exact expected version.

However, due to a bug, even `--enable-perlinterop=dynamic` would let
`vim.exe` link to the Perl DLL, defeating the purpose of the `dynamic`
part.

This was fixed in the VIM project, via
https://github.com/vim/vim/commit/55460da26c27, and here is a backport
to get this fix into MSYS2.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>