msys2-runtime-3.5: branch off of msys2-runtime
Cygwin v3.5.0 has been released end of January 2024, and we were _about_
to be on the brink to switch to MSYS2 runtime based on the Cygwin
runtime v3.5.x.

However:

- That would leave Windows 7/8 compatibility behind.
- There have been a couple of hiccups in our testing, which led us to
  skip v3.5.0 already.
- We determined that a safer way would be to let users opt into using
  v3.5.x first, kicking the tires, so to say.

So let's establish a `msys2-runtime-3.5` package (much like
`msys2-runtime-3.3`) to be able to easily revert to a known state really
quickly, including the ability to run on Windows 7.

For now, this is merely a verbatim copy of the files in `msys2-runtime`,
and the next commits will modify them to reflect the version bump.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>