bash: Fix CR handling
Fixes #1839

`0005-bash-4.3-msys2-fix-lineendings.patch` adds CRLF support.
However, `0001-bash-4.4-cygwin.patch` already added `igncr` option to
Bash to support CRLF.

I confirmed that the Cygwin version of Bash has also the same issue
with #1839 when the `igncr` option is set.

After debugging, I found that there is an issue in
`rewind_input_string()` in `parser.y` that it doesn't take the CR into
account.

This PR adds the following changes:
* Modify `rewind_input_string()` to take the CR into account.
  (It might be better to apply a similar change to the Cygwin version of
  Bash.)
* Remove all the changes from `y.tab.c`. This file should be
  automatically generated from `parser.y`.  (Also, added bison to the
  make dependencies.)
* Set `LC_ALL=C.UTF-8` when running `make check` for running on
  non-English locales.
* Add two tests: `run-ps1lf` and `run-crlf`.

Note: This patch contains a line with CRLF. So, .gitattributes is also
updated to keep the CRLF.