msys2-runtime-3.4(update-patches.sh): do generate stable checksums for the Cygwin code
As of v6.1.0, Pacman no longer generates `SKIP` as SHA-256 checksums for
Git repositories, as per the release notes at
https://gitlab.archlinux.org/pacman/pacman/-/blob/v6.1.0/NEWS?ref_type=tags#L21:

> - Add checksum support for git/mercurial/bzr sources

Instead of outputting `SKIP`, Pacman now runs `git -c core.abbrev=no
archive --format tar <branch-or-tag> | sha256sum` and uses the result:
https://gitlab.archlinux.org/pacman/pacman/-/commit/2fc2ab6cf0fbb93e1b3182a1997d3c9ffc9fc0fd

Unfortunately, the output of `git archive` is not stable. For one, the
`core.autoCRLF` setting (which defaults to `true` in Git for Windows)
affects the result.

Let's ensure that that `core.autoCRLF` setting is forced to `false`,
with the idea to add other settings as required.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>