ncurses: enable term-driver for TERM='#win32con'
Add a patch to enable the term-driver support for the Win32 console with
TERM='#win32con', which does not require a terminfo database.

Does not change the regular terminfo-based functionality of ncurses
otherwise.

This patch has been submitted upstream and this is the rebased version
for 6.5.

You can verify that this works by running an ncurses program with
TERM='#win32con' in a Windows Terminal session or similar. For example
this one:

https://github.com/mtoyoda/sl
.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>