ncurses: update to 6.5-20240831,enable term-driver
Update to the current dev version, as the last stable release was
already two years ago.

This patch set includes a patch to enable the term-driver support for
the Win32 console with TERM='#win32con' on the Cygwin runtime, like it
does for MinGW, which does not require a terminfo database.

This does not change the regular terminfo-based functionality of ncurses
otherwise.

You can verify that this works by running an ncurses program with
TERM='#win32con' in a Windows Terminal session or similar. For example
this one:

https://github.com/mtoyoda/sl
.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>