perl: build for cygwin
The goal is for perl to report that it was built for cygwin
($^O gives "cygwin" for example) and to adjust all the code depending
on that and remove all the patches for handing the "msys" case.
Everything else should work as before.

* In case cygwin check was extended with msys, remove msys case
* In case msys check was added, replace with cygwin check
* In case msys specific checks/code was added, merge that into the cygwin paths

We still check/set the MSYS env var instead of the CYGWIN env var,
since msys2-runtime doesn't handle both yet, so keep that as is for now.