perl: remove renames in comments
This is since this used to be a separate msys.sh file
which was synced back.

Let's keep things simple and only patch what we need to.

TODO: check if the other changes in that file are really needed,
or if they just diverged since the split for no good reason.