filesystem: Remove /etc/bash.bash_logout
By adding `Hold=always` in ~/.minttyrc, it is possible to keep MinTTY open
after all processes in it have terminated, which is useful for examination
of outputs. However this script used to clear the screen, leaving a blank
window, and causing confusion.

Another reason for removing this script is that it's not found on fresh
installation of Debian or Ubuntu.

Reference: https://superuser.com/questions/1262655/mintty-how-to-keep-my-output-visible-after-processes-terminate/1263001#1263001
Signed-off-by: LIU Hao <lh_mouse@126.com>