genromfs: close() fds after use
The current version of genromfs has a bug whereby it leaks fd handles by not
closing them after use. This results in the tool failing with large sets of
files (approximately 9,000) displaying an error: "Too many open files".

Linux users are typically less affected by this bug because the operating
system in most cases allows a far greater number of simultaneously open file
handles than is allowed on Windows.

A patch to correct the issue has been provided by @NomanKhan777 here:

  https://github.com/chexum/genromfs/pull/5

However, the maintainer does not appear to be very responsive. The most recent
commit in the repository was made in 2011. Unless and until this patch is
accepted upstream, or a new maintainer is found for the project, this commit
adds the patch to MSYS2-packages to correct the issue.

Signed-off-by: Joel Holdsworth <jholdsworth@nvidia.com>