libassuan: work around DLL hell
Currently, `gpg.exe` links to `msys-assuan-0.dll`. Which means that
after installing the new `libassuan` version, no `gpg.exe` invocation will
succeed, including the one needed to verify the integrity of the files
needed to build a new `gnupg` package... Catch-22.

To work around that, simply copy the new DLL to the old name; This seems
to work well enough, and we will need to drop that work-around once both
packages that depend on `libassuan` (`gnupg` and `pinentry`) are
rebuilt.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>