zsh: fix performance issues related to PATH
When $PATH contains "/" and command "foo" is run, zsh will try to execute "//foo"
instead of "/foo", which will attempt to contact a network share on server "foo",
causing DNS and CIFS requests that will greatly slow down the shell. And if "/foo"
actually exists, it will not be found.

As a workaround, we do not insert the additional '/' if the path element already
ends with a slash.

Another workaround for the users would be to put "/." instead of "/" in $PATH.