msys2-runtime: console: tty::restore really restores the previous mode
Previously, tty::restore sets the console mode to a predetermined
console mode that is widely common for many non-cygwin console apps.
So, if a non-cygwin app that is started from cygwin process changes
the console mode and executes cygwin sub-process, the console mode
is changed to the predetermined mode rather than being restored the
original mode that is set by the non-cygwin app.
With this patch, the console mode is stored when a cygwin process is
started from non-cygwin app, then tty::restore restores the previous
console mode that is used by the previous non-cygwin app when the
cygwin app exits.

This is a companion of https://github.com/msys2/msys2-runtime/pull/269.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>