makepkg confs: apply +crt-static in RUSTFALGS for CLANG*
currently there is an issue with `crt-static` option being
uncontrollable for windows-gnu targets: . this behavour is bad because
libunwind.dll is linked by default for windows-gnullvm, but no gcc
dynamic libs are linked in case of windows-gnu. I tried to use
bootstrap.toml option, but it doesn't do what is expected.

this change should be temp before the issue is fixed. this flag can be
changed in case of package which has cc-libs as their dependency anyway.

I'll block some rust updates/additions before this is merged