itstool: use proposed patch to port to python-lxml
So we can get rid of libxml2-python, which is deprecated upstream.
It's not clear when upstream will integrate it so use the porposed
patch for now.

Remove patches which conflict with the above, with the hope that
moving to lxml might have made the obsolete, or the issues are no
longer there.

If there are any regressions, please speak up and we'll have a look.