curl: drop heimdal support
Upstream dropped support for it:
https://github.com/curl/curl/commit/8be9a26451d466571864a608bc48cc77746ff00d

The alternative is krb5, but that has no cygwin support upstream.
There is an orphaned cygwin package with 45 patches
https://cygwin.com/cgit/cygwin-packages/krb5
which I don't want to maintain really.

So let's drop heimdal support now, so if someone screams we can try to
revert and look into it more.