pacman-mirrors: remove env specific mirror lists (#5886)
These were added in https://github.com/msys2/MSYS2-packages/pull/2569
since we switched to a unified list by default in
https://github.com/msys2/MSYS2-packages/pull/2572

They were kept in case someone had modified their pacman config and
not keep it in sync.

It's been 4.5 years now, so remove those files. While they don't hurt,
they can be confusing to users as they are in the same dir but just not
used, and changes to them are thus ignored.